+++ BE/house/src/main/java/com/nextsquad/house/dto/login/KakaoAccessTokenResponseDto.java
@@ -1,10 +1,10 @@
-package com.nextsquad.house.dto;
+package com.nextsquad.house.dto.login;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
 @Getter
-public class NaverAccessTokenResponseDto {
+public class KakaoAccessTokenResponseDto {
     @JsonProperty("access_token")
     private String accessToken;
 }
