+++ BE/house/src/main/java/com/nextsquad/house/dto/rentarticle/HouseFacilityListDto.java
@@ -1,4 +1,4 @@
-package com.nextsquad.house.dto;
+package com.nextsquad.house.dto.rentarticle;
 
 import com.nextsquad.house.domain.house.HouseFacility;
 import lombok.AllArgsConstructor;
