+++ BE/house/src/main/java/com/nextsquad/house/dto/user/UserResponseDto.java
@@ -1,4 +1,4 @@
-package com.nextsquad.house.dto;
+package com.nextsquad.house.dto.user;
 
 import com.nextsquad.house.domain.user.User;
 import lombok.AllArgsConstructor;
