+++ BE/house/src/main/java/com/nextsquad/house/repository/rentarticle/CustomRentArticleRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.nextsquad.house.repository;
+package com.nextsquad.house.repository.rentarticle;
 
 import com.nextsquad.house.domain.house.RentArticle;
 import com.nextsquad.house.dto.SearchConditionDto;
