+++ BE/house/src/main/java/com/nextsquad/house/repository/wantedarticle/WantedArticleRepository.java
@@ -1,4 +1,4 @@
-package com.nextsquad.house.repository;
+package com.nextsquad.house.repository.wantedarticle;
 
 import com.nextsquad.house.domain.house.WantedArticle;
 import com.nextsquad.house.domain.user.User;
