+++ BE/house/src/main/java/com/nextsquad/house/service/LoginService.java
@@ -2,8 +2,8 @@ package com.nextsquad.house.service;
 
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.nextsquad.house.domain.user.User;
-import com.nextsquad.house.dto.JwtResponseDto;
-import com.nextsquad.house.dto.OauthLoginRequestDto;
+import com.nextsquad.house.dto.login.JwtResponseDto;
+import com.nextsquad.house.dto.login.OauthLoginRequestDto;
 import com.nextsquad.house.exception.OauthClientNotFoundException;
 import com.nextsquad.house.exception.UserNotFoundException;
 import com.nextsquad.house.login.jwt.JwtProvider;
@@ -11,7 +11,7 @@ import com.nextsquad.house.login.jwt.JwtToken;
 import com.nextsquad.house.login.oauth.OauthClient;
 import com.nextsquad.house.login.oauth.OauthClientMapper;
 import com.nextsquad.house.login.userinfo.UserInfo;
-import com.nextsquad.house.repository.UserRepository;
+import com.nextsquad.house.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
