+++ BE/house/src/main/java/com/nextsquad/house/service/RentArticleService.java
@@ -5,9 +5,14 @@ import com.nextsquad.house.domain.house.*;
 import com.nextsquad.house.domain.user.User;
 import com.nextsquad.house.dto.*;
 import com.nextsquad.house.dto.bookmark.BookmarkRequestDto;
+import com.nextsquad.house.dto.rentarticle.*;
 import com.nextsquad.house.exception.*;
 import com.nextsquad.house.login.jwt.JwtProvider;
-import com.nextsquad.house.repository.*;
+import com.nextsquad.house.repository.rentarticle.HouseFacilityRepository;
+import com.nextsquad.house.repository.rentarticle.HouseImageRepository;
+import com.nextsquad.house.repository.rentarticle.RentArticleBookmarkRepository;
+import com.nextsquad.house.repository.rentarticle.RentArticleRepository;
+import com.nextsquad.house.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
