+++ BE/house/src/main/java/com/nextsquad/house/service/UserService.java
@@ -6,11 +6,19 @@ import com.nextsquad.house.domain.house.WantedArticle;
 import com.nextsquad.house.domain.house.WantedArticleBookmark;
 import com.nextsquad.house.domain.user.User;
 import com.nextsquad.house.dto.*;
+import com.nextsquad.house.dto.rentarticle.RentArticleListElement;
+import com.nextsquad.house.dto.rentarticle.RentArticleListResponse;
 import com.nextsquad.house.dto.user.DuplicationCheckResponse;
+import com.nextsquad.house.dto.user.UserInfoDto;
+import com.nextsquad.house.dto.user.UserResponseDto;
 import com.nextsquad.house.dto.wantedArticle.WantedArticleElementResponse;
 import com.nextsquad.house.dto.wantedArticle.WantedArticleListResponse;
 import com.nextsquad.house.exception.UserNotFoundException;
-import com.nextsquad.house.repository.*;
+import com.nextsquad.house.repository.rentarticle.RentArticleBookmarkRepository;
+import com.nextsquad.house.repository.rentarticle.RentArticleRepository;
+import com.nextsquad.house.repository.user.UserRepository;
+import com.nextsquad.house.repository.wantedarticle.WantedArticleBookmarkRepository;
+import com.nextsquad.house.repository.wantedarticle.WantedArticleRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
