+++ BE/house/src/test/java/com/nextsquad/house/acceptance/login/LoginAcceptanceTest.java
@@ -2,13 +2,13 @@ package com.nextsquad.house.acceptance.login;
 
 import com.nextsquad.house.config.RestDocsConfiguration;
 import com.nextsquad.house.domain.user.User;
-import com.nextsquad.house.dto.OauthLoginRequestDto;
+import com.nextsquad.house.dto.login.OauthLoginRequestDto;
 import com.nextsquad.house.exception.UserNotFoundException;
 import com.nextsquad.house.login.jwt.JwtProvider;
 import com.nextsquad.house.login.jwt.JwtToken;
 import com.nextsquad.house.login.oauth.*;
 import com.nextsquad.house.login.userinfo.UserInfo;
-import com.nextsquad.house.repository.UserRepository;
+import com.nextsquad.house.repository.user.UserRepository;
 import com.nextsquad.house.service.RedisService;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
