+++ BE/house/src/test/java/com/nextsquad/house/acceptance/rentArticle/RentArticleAcceptanceTest.java
@@ -1,11 +1,11 @@
 package com.nextsquad.house.acceptance.rentArticle;
 
-import com.nextsquad.house.dto.HouseFacilityListDto;
-import com.nextsquad.house.dto.RentArticleRequest;
+import com.nextsquad.house.dto.rentarticle.HouseFacilityListDto;
+import com.nextsquad.house.dto.rentarticle.RentArticleRequest;
 import com.nextsquad.house.dto.bookmark.BookmarkRequestDto;
 import com.nextsquad.house.login.jwt.JwtProvider;
 import com.nextsquad.house.login.jwt.JwtToken;
-import com.nextsquad.house.repository.UserRepository;
+import com.nextsquad.house.repository.user.UserRepository;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
