+++ BE/house/src/test/java/com/nextsquad/house/domain/house/RentArticleTest.java
@@ -1,8 +1,7 @@
 package com.nextsquad.house.domain.house;
 
 import com.nextsquad.house.domain.user.User;
-import com.nextsquad.house.dto.RentArticleRequest;
-import org.assertj.core.api.Assertions;
+import com.nextsquad.house.dto.rentarticle.RentArticleRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +11,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class RentArticleTest {
