+++ BE/house/src/test/java/com/nextsquad/house/repository/RentArticleBookmarkRepositoryTest.java
@@ -3,6 +3,9 @@ package com.nextsquad.house.repository;
 import com.nextsquad.house.domain.house.RentArticle;
 import com.nextsquad.house.domain.house.RentArticleBookmark;
 import com.nextsquad.house.domain.user.User;
+import com.nextsquad.house.repository.rentarticle.RentArticleBookmarkRepository;
+import com.nextsquad.house.repository.rentarticle.RentArticleRepository;
+import com.nextsquad.house.repository.user.UserRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
