+++ BE/house/src/test/java/com/nextsquad/house/repository/WantedArticleBookmarkRepositoryTest.java
@@ -3,6 +3,9 @@ package com.nextsquad.house.repository;
 import com.nextsquad.house.domain.house.WantedArticle;
 import com.nextsquad.house.domain.house.WantedArticleBookmark;
 import com.nextsquad.house.domain.user.User;
+import com.nextsquad.house.repository.user.UserRepository;
+import com.nextsquad.house.repository.wantedarticle.WantedArticleBookmarkRepository;
+import com.nextsquad.house.repository.wantedarticle.WantedArticleRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
