+++ BE/house/src/test/java/com/nextsquad/house/repository/WantedArticleRepositoryTest.java
@@ -4,6 +4,8 @@ import com.nextsquad.house.domain.house.WantedArticle;
 import com.nextsquad.house.domain.user.User;
 import com.nextsquad.house.dto.SearchConditionDto;
 import com.nextsquad.house.login.oauth.OauthClientType;
+import com.nextsquad.house.repository.user.UserRepository;
+import com.nextsquad.house.repository.wantedarticle.WantedArticleRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
